LLM_project/pyproject.toml
[project]
name = "data-analysis-agent"
version = "0.1.0"
description = "Data Analysis Agent with FastAPI backend and Streamlit frontend"
requires-python = ">=3.9"


dependencies = [
    # Core Application
    "fastapi==0.115.12",
    "uvicorn==0.34.0",
    "python-multipart==0.0.6",
    "pydantic==2.10.6",
    "python-dotenv==1.0.1",
    
    # AI/ML Core
    "langchain==0.3.21",
    "langchain-community==0.3.20",
    "langchain-core==0.3.49",
    "ollama==0.4.7",
    "langchain-ollama==0.3.0",
    
    # Utilities
    "loguru==0.7.2",
    "pyzmq==25.1.1",
    "python-dateutil==2.8.2",
    "tqdm==4.66.1",
    "pyyaml==6.0.1",
    "toml==0.10.2",
    "jsonschema==4.20.0",
    "requests==2.31.0",
]

[project.optional-dependencies]
data = [
    "pandas==2.1.3",
    "numpy==1.26.3",
    "openpyxl==3.1.2",
    "pyarrow==18.0.0",
    "scipy==1.13.1",
    "numexpr==2.10.2",
]

viz = [
    "matplotlib==3.10.1",
    "seaborn==0.13.2",
    "plotly==5.24.1",
    "pillow==10.0.0",
]

web = [
    "streamlit==1.39.0",
]

test = [
    "pytest==8.3.3",
    "httpx==0.28.1",
    "pytest-asyncio==0.23.5",
]

docs = [
    "mkdocs>=1.0",
    "mkdocstrings[python]>=0.20",
    "mkdocs-material>=9.6.4",
    "pymdown-extensions",
    "pygments",
]


[tool.ruff]
line-length = 102
select = ["ALL"]
